requirements.txt
ccxt>=4.5.0
pandas>=2.2.0
numpy>=1.24.3
torch>=2.2.0
scikit-learn>=1.4.1
ta>=0.10.2
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
flasgger>=0.9.7
